extracted default layout
